feat: moving DeferRulesList code to its own file and directory
