+++ src/pages/RuleSheet/Defer/screens/DeferRulesList/index.ts
@@ -1,0 +1,1 @@
+export * from './DeferRulesList';
